+++ cli/src/commands/mod.rs
@@ -1,9 +1,11 @@
 mod app;
 mod check;
 mod preset;
+mod self_install;
 mod shell;
 
 pub(crate) use app::*;
 pub(crate) use check::*;
 pub(crate) use preset::*;
+pub(crate) use self_install::*;
 pub(crate) use shell::*;